+++ src/carma2/common/world.h
@@ -72,7 +72,7 @@ void C2_HOOK_FASTCALL SetCarSimplificationLevel(int pLevel);
 
 int C2_HOOK_FASTCALL GetCarSimplificationLevel(void);
 
-int C2_HOOK_CDECL ProcessMaterials(br_actor* pActor, material_cbfn* pCallback);
+intptr_t C2_HOOK_CDECL ProcessMaterials(br_actor* pActor, material_cbfn* pCallback);
 
 intptr_t C2_HOOK_CDECL SetAccessoryRenderingCB(br_actor* pActor, void* pFlag);
